% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/courseraquakes-package.R
\docType{package}
\name{courseraquakes-package}
\alias{courseraquakes}
\alias{courseraquakes-package}
\title{courseraquakes: Clean Up Earthquake Data from NCIE}
\description{
This package is final project for coursera course. It cleans up data and create timeline plot.
}
\author{
\strong{Maintainer}: Ibrahim Nadiyan \email{ibrahimnadian@yahoo.co.id}

}
\keyword{internal}
