% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacleaning.R
\docType{data}
\name{earthquake_raw}
\alias{earthquake_raw}
\title{National Centers for Environmental Information earthquake Data}
\format{
' A data frame with 6404 rows and 39 columns:
\describe{
\item{Search Parameters}{Lorem ipsum}
\item{Year}{Lorem ipsum}
\item{Mo}{Lorem ipsum}
\item{Dy}{Lorem ipsum}
\item{Hr}{Lorem ipsum}
\item{Mn}{Lorem ipsum}
\item{Sec}{Lorem ipsum}
\item{Tsu}{Lorem ipsum}
\item{Vol}{Lorem ipsum}
\item{Location Name}{Lorem ipsum}
\item{Latitude}{Lorem ipsum}
\item{Longitude}{Lorem ipsum}
\item{Focal Depth (km)}{Lorem ipsum}
\item{Mag}{Lorem ipsum}
\item{MMI Int}{Lorem ipsum}
\item{Deaths}{Lorem ipsum}
\item{Death Description}{Lorem ipsum}
\item{Missing}{Lorem ipsum}
\item{Missing Description}{Lorem ipsum}
\item{Injuries}{Lorem ipsum}
\item{Injuries Description}{Lorem ipsum}
\item{Damage ($Mil)}{Lorem ipsum}
\item{Damage Description}{Lorem ipsum}
\item{Houses Destroyed}{Lorem ipsum}
\item{Houses Destroyed Description}{Lorem ipsum}
\item{Houses Damaged}{Lorem ipsum}
\item{Houses Damaged Description}{Lorem ipsum}
\item{Total Deaths}{Lorem ipsum}
\item{Total Death Description}{Lorem ipsum}
\item{Total Missing}{Lorem ipsum}
\item{Total Missing Description}{Lorem ipsum}
\item{Total Injuries}{Lorem ipsum}
\item{Total Injuries Description}{Lorem ipsum}
\item{Total Damage ($Mil)}{Lorem ipsum}
\item{Total Damage Description}{Lorem ipsum}
\item{Total Houses Destroyed}{Lorem ipsum}
\item{Total Houses Destroyed Description}{Lorem ipsum}
\item{Total Houses Damaged}{Lorem ipsum}
\item{Total Houses Damaged Description}{Lorem ipsum}
}
}
\source{
\url{https://www.ngdc.noaa.gov/hazel/view/hazards/earthquake/event-data?maxYear=2023&minYear=-2150}
}
\usage{
earthquake_raw
}
\description{
Raw Earthquake data recorded by NCIE from -2150 to 2023
}
\keyword{datasets}
