% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacleaning.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{Helper function for initial data cleaning}
\usage{
eq_location_clean(locname)
}
\arguments{
\item{locname}{String from raw earthquake data with country data}
}
\value{
Country name if the input has no ":", otherwise any string after the country
}
\description{
Helper function for initial data cleaning
}
\keyword{internal}
