% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eq_map.R
\name{eq_map}
\alias{eq_map}
\title{Earthquake interactive map builder}
\usage{
eq_map(df, annotation = "popup_text", ...)
}
\arguments{
\item{df}{cleaned data frame, preferably one that has been filtered}

\item{annotation}{Character string containing either NULL, "popup_text", or column name}

\item{...}{
  Arguments passed on to \code{\link[=datecountry_filter]{datecountry_filter}}
  \describe{
    \item{\code{minyear,maxyear}}{Integer value of minimum and maximum year, accept any number}
    \item{\code{minmonth,maxmonth}}{Integer value of minimum and maximum month, only accept 1-12}
    \item{\code{minday,maxday}}{Integer value of minimum and maximum date, make sure it is valid date}
    \item{\code{countries}}{Character vector containing country name in all upper case}
  }}
}
\value{
interactive map made by leaflet
}
\description{
A wrapper for leaflet package that shows interactive map containing marker.
The marker might or might not have popup information. If annotation = NULL, return
interactive map without any popup. If annotation = 'popup_text', put location, magnitude,
and total death into popup information. Any missing information will be removed.
Alternatively, you can specify specific column.
}
\examples{
clean_df <- initial_cleaning()
eq_map(clean_df, annotation = "popup_text",
        minyear = 2012, maxyear = 2022, countries = c('IRAN', 'IRAQ'))
}
\seealso{
\code{\link[=datecountry_filter]{datecountry_filter()}}
}
