% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_label.R
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\alias{geom_timeline_label}
\title{Timeline label}
\format{
An object of class \code{GeomTimelineLabel} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 7.
}
\usage{
GeomTimelineLabel

geom_timeline_label(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  parse = FALSE,
  check_overlap = FALSE,
  size.unit = "mm",
  na.rm = TRUE,
  show.legend = NA,
  inherit.aes = TRUE,
  nmax = NULL
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{aes()}}. If specified and
\code{inherit.aes = TRUE} (the default), it is combined with the default mapping
at the top level of the plot. You must supply \code{mapping} if there is no plot
mapping.}

\item{data}{The data to be displayed in this layer. There are three
options:

If \code{NULL}, the default, the data is inherited from the plot
data as specified in the call to \code{\link[ggplot2:ggplot]{ggplot()}}.

A \code{data.frame}, or other object, will override the plot
data. All objects will be fortified to produce a data frame. See
\code{\link[ggplot2:fortify]{fortify()}} for which variables will be created.

A \code{function} will be called with a single argument,
the plot data. The return value must be a \code{data.frame}, and
will be used as the layer data. A \code{function} can be created
from a \code{formula} (e.g. \code{~ head(.x, 10)}).}

\item{stat}{The statistical transformation to use on the data for this
layer, either as a \code{ggproto} \code{Geom} subclass or as a string naming the
stat stripped of the \code{stat_} prefix (e.g. \code{"count"} rather than
\code{"stat_count"})}

\item{position}{Position adjustment, either as a string naming the adjustment
(e.g. \code{"jitter"} to use \code{position_jitter}), or the result of a call to a
position adjustment function. Use the latter if you need to change the
settings of the adjustment.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.
It can also be a named logical vector to finely select the aesthetics to
display.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link[ggplot2:borders]{borders()}}.}

\item{nmax}{Optional, an integer expressing the amount of label you want to show}

\item{Magnitude}{Optional, a column that will be used to sort the data frame}
}
\value{
ggplot2 label layer
}
\description{
\code{GeomTimelineLabel} and \code{geom_timeline_label} are ggproto class and function
that is used to add label to geom_timeline. If you don't want to show every single label,
use nmax to specify how many labels you want to show. If you use nmax, you have to specify
Magnitude so the geom can sort the data frame and pick nmax number of row from the top.
The geom will not issue warning if there is removed data from NA value, as it's the way the geom
sort and filter the data. The user have to make sure their data contain no unintended NA value.
}
\seealso{
\code{\link[=timeline]{timeline()}}

\code{\link[=geom_timeline]{geom_timeline()}}
}
\keyword{datasets}
