% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datacleaning.R
\name{initial_cleaning}
\alias{initial_cleaning}
\title{Data cleaning for raw earthquake data recorded by NCIE from -2150 to 2023.}
\usage{
initial_cleaning(df = earthquake_raw)
}
\arguments{
\item{df}{Tibble data frame taken from NCIE, if you have different data, you can supply your own.}
}
\value{
Cleaned tibble data frame with separate country and area column, along with improved date stamp
}
\description{
A function to clean up NCIE earthquake data, this includes replacing missing value,
changing non-descriptive column name, separating country from its area, and creating
column with date class.
}
\examples{
clean_df <- initial_cleaning()
}
