% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_and_timeline.R
\name{timeline}
\alias{timeline}
\title{\verb{Timeline construction}}
\usage{
timeline(df, y = NULL, nmax = 0, label = FALSE, ...)
}
\arguments{
\item{df}{Data frame or its extension containing cleaned up earthquake data}

\item{y}{Column name that will be used as the basis of timeline separation}

\item{nmax}{Optional, an integer expressing the amount of label you want to show}

\item{label}{Boolean value, whether you want to show label containing location name or not.}

\item{...}{
  Arguments passed on to \code{\link[=datecountry_filter]{datecountry_filter}}
  \describe{
    \item{\code{minyear,maxyear}}{Integer value of minimum and maximum year, accept any number}
    \item{\code{minmonth,maxmonth}}{Integer value of minimum and maximum month, only accept 1-12}
    \item{\code{minday,maxday}}{Integer value of minimum and maximum date, make sure it is valid date}
    \item{\code{countries}}{Character vector containing country name in all upper case}
  }}
}
\value{
ggplot2 plot
}
\description{
This is wrapper function that filter out the data frame and then pass it into
custom \code{geom_timeline} and (optionally) \code{geom_timeline_label} The theme and label are
already chosen, if you wish to customize those, you might  want to construct your own
ggplot2 using \code{geom_timeline} and \code{geom_timeline_label}
}
\examples{
clean_df <- initial_cleaning()
timeline(data = clean_df, y = Country, nmax = 10, label = TRUE,
                              minyear =  2012, maxyear = 2021, countries = c('ARGENTINA', 'PALESTINE'))
}
